==> 3-kanji.ipynb <==
from module import Example

import os
import glob
import pandas as pd
import numpy as np
import re
import matplotlib.pyplot as plt
import functools as ft
import operator as op
import unicodedata

os.chdir('/Users/sheg/Documents/projects/anki-flashcards/japanese/data')

# Загружаем Радикалы

У оригинального файла `japanese-radicals.csv` отсутствовало чтение для радикала 乙, оно добавлено вручную при редактировании этого файла

radic = pd.read_csv('japanese-radicals.csv', delimiter=';',
                    usecols=[0, 1, 3, 4, 6, 8],
                    header=0,
                    names=['strokes', 'radic', 'en', 'reading', 'position', 'important'],
                    dtype={'strokes': 'Int8'}
                   )

radic.important.replace({np.nan: False, 'Important': True}, inplace=True)
radic.reset_index(drop=True, inplace=True)

from lxml import html, etree

from requests import get

page = get('https://en.wikipedia.org/wiki/List_of_kanji_radicals_by_stroke_count')

xpath = etree.XPath('//*[@id="mw-content-text"]/div/table[1]')
table = xpath( html.fromstring(page.text) )[0]

table = pd.read_html( etree.tostring(table), index_col=0 )[0]

kangxi = table.iloc[:, 0]
kangxi.Name = 'radic'

kangxi = kangxi.str.rstrip('\)').str.split(' \(')

f = lambda x: {'radic': x[0], 'var': ','.join(x[1:]) if len(x) - 1 else np.nan}
kangxi = kangxi.apply(f)

kangxi = pd.DataFrame( list(kangxi) )

# normalization
kangxi = kangxi.applymap(
    lambda rad: unicodedata.normalize('NFKC', rad) if isinstance(rad, str) else rad
)

radic.radic = radic.radic.apply(lambda rad: unicodedata.normalize('NFKC', rad))

radic = kangxi.merge(radic, on='radic', how='left')

radic.drop_duplicates('radic', inplace=True)

radic.index = range(1, 215)

radic.to_pickle('radic.pkl')

# Кандзи

## Данные KANJIDIC

from lxml import etree

kanjidic = etree.parse('kanjidic2.xml').getroot()

kanjidic = kanjidic.xpath('character')

xpaths = {
    'kanji': 'literal',
    'radic': 'radical/rad_value',
    'grade': 'misc/grade',
    'strokes': 'misc/stroke_count',
    'freq_kd': 'misc/freq',
    'jlpt': 'misc/jlpt',
    'on_kd': "reading_meaning/rmgroup/reading[@r_type='ja_on']",
    'kun_kd': "reading_meaning/rmgroup/reading[@r_type='ja_kun']",
    'en_kd': 'reading_meaning/rmgroup/meaning[not(@m_lang)]'
}

df = pd.DataFrame(columns=xpaths.keys())

for kanji in kanjidic:
    row = {col: kanji.xpath(xpath_ + '/text()') for col, xpath_ in xpaths.items()}
    df = df.append(row, ignore_index=True)

df.kanji = df.kanji.apply(lambda l: l[0])

df.grade = df.grade.apply(lambda l: l[0] if len(l) == 1 else np.nan)

df.grade = df.grade.astype(float)

# Первая ступень школы
df['kyoiku'] = df.apply(lambda row: row.grade if row.grade in {1, 2, 3, 4, 5, 6} else np.nan, axis=1).astype('Int8').astype('category')

# Для повседневного употребления, кандзи, не входящие в этот список, обычно подписываются фуриганой
df['joyo'] = df.apply(lambda row: True if row.grade in {1, 2, 3, 4, 5, 6, 8} else False, axis=1)

# Для использования в именах
df['jinmeiyo'] = df.apply(lambda row: True if row.grade in {1, 2, 3, 4, 5, 6, 8, 9, 10} else False, axis=1)

# Больше не нужна
del df['grade']

df.freq_kd = df.freq_kd.apply(lambda l: l[0] if len(l) == 1 else np.nan).astype(float).astype('Int32')

df.jlpt = df.jlpt.apply(lambda l: l[0] if len(l) == 1 else np.nan).astype(float).astype('Int8')

df.radic = df.radic.apply(lambda l: [int(n) for n in l])

df.strokes = df.strokes.apply(lambda l: [int(n) for n in l])

# Переназовём таблицу
kanji = df
del df

kanji.head()

## Добавим переводы кандзи на русский, взятые на [nippon.temerov.org]

temerov_files = glob.glob('kyoiku/*.*sv') + ['joyo.tsv']

df = pd.DataFrame(columns=['kanji', 'ru_tm'])

for file in temerov_files:
    tmp = pd.read_csv(file, delimiter='\t',
                      usecols=[0, 1], header=0,
                      names=df.columns)
    
    df = df.append(tmp, ignore_index=True)

kanji = kanji.merge(df, on='kanji', how='outer')

del df, tmp

kanji.head()

## Добавим переводы на английский, он, кун и примеры из проекта *KanjiAlive*

def parse_ka_examples(s):
    
    l = eval(s)
    examples = []
    
    for jp, en in l:
        kanji, kana = jp.rsplit('（', 1)
        kana = kana.rstrip('）')

        ex = Example(kanji=kanji, kana=kana, en=en)
        examples.append(ex)
    
    return examples

df = pd.read_csv('ka_data.csv', usecols=[0, 3, 5, 7, 9], header=0,
                names=['kanji', 'en_ka', 'kun_ka', 'on_ka', 'examples_ka'])

df.examples_ka = df.examples_ka.apply(parse_ka_examples)

df.kun_ka = df.kun_ka.str.split('、')
df.on_ka = df.on_ka.str.split('、')

kanji = kanji.merge(df, on='kanji', how='outer')

kanji.head()

kanji.to_pickle('kanji_full.pkl')
del kanji

## Последние штрихи

kanji_full = pd.read_pickle('kanji_full.pkl')

### Темеров

Приведём в порядок, для начала разрежем переводы-строки Темерова на списки

# Точка используются в двух переводах, её трогать не будем
kanji_full[ kanji_full.ru_tm.str.contains('[.]') == True ]

# Запятые используются для разделения толкований, разрежем по запятым, а остальные переводы запакуем в списки
kanji_full.ru_tm = kanji_full.ru_tm.str.split(',')

### Разрежем переводы KanjiAlive

# Аналогично, точки используются лишь в многоточиях (одна карточка), запятые разграничивают толкование
kanji_full[ kanji_full.en_ka.str.contains('[.]') == True ]

kanji_full.en_ka = kanji_full.en_ka.str.split(',')

### Заменим пустые списки `[]` в колонках на `np.nan`

for col in kanji_full.columns:
    kanji_full[col][ kanji_full[col].apply(lambda x: x == []) == True ] = np.nan

### Заменим float NaN на np.nan

kanji_full.fillna(np.nan, inplace=True)

### Нормализуем символы

kanji_full.kanji = kanji_full.kanji.apply(lambda s: unicodedata.normalize('NFKC', s))

kanji_full.to_pickle('kanji_full.pkl')

# Выберем нужное подмножество данных

kanji_full = pd.read_pickle('kanji_full.pkl')

get_rules = {
    'kanji': 'kanji',
    'radic': 'radic',
    'strokes': 'strokes',
    'freq': 'freq_kd',
    'jlpt': 'jlpt',
    'on': lambda x: x.on_ka if x.on_ka is not np.nan else x.on_kd,
    'kun': lambda x: x.kun_ka if x.kun_ka is not np.nan else x.kun_kd,
    'en_kd': 'en_kd',
    'en_ka': 'en_ka',
    'ru': 'ru_tm',
    'examples': 'examples_ka',
    'kyoiku': 'kyoiku',
    'joyo': 'joyo',
    'jinmeiyo': 'jinmeiyo'
}

kanji = pd.DataFrame()

for col_name, rule in get_rules.items():

    if type(rule) is str:
        col = kanji_full[rule]
    else:
        col = kanji_full.apply(rule, axis=1)
    
    kanji[col_name] = col

kanji = kanji.sort_values('freq')

kanji.to_pickle('kanji.pkl')
