==> 5-search.ipynb <==
import os
import glob
import ray
import pandas as pd
import numpy as np
import re
import matplotlib.pyplot as plt
import functools as ft
import operator as op
from module import JP_Parser, Example

os.chdir('/Users/sheg/Documents/projects/anki-flashcards/japanese/data')


# Поиск примеров

jp_morpho = pd.read_pickle('jp_morpho.pkl')

sent = pd.read_pickle('sent.pkl')

def jp_examples(text):
    katakana = JP_Parser.reading(text, 'katakana')
    word_matches = jp_morpho[jp_morpho.reading == katakana]
    sent_matches_ids = word_matches.index.get_level_values(0)
    
    return sent.loc[sent_matches_ids]

jp_examples('日本語').head()
